requirements.txt
pyserial
numpy
Pillow
opencv-python 